Introduce separate BaseModel for list responses (#33)
* Introduce separate BaseModel for list responses

Mypy couldn't properly typehint our function responses when list values
were passed, because our main SchemaType variable is pinned to a
BaseModel `base` value. Lists would work at runtime but not during
typehinting, which made it harder to debug client code.

Here we standardize the generic API so everything passed in is a
BaseModel. We also add a small helper function `ListResponse` that
allows clients to easily request a list as output. Because this
ListResponse is typehinted generically, we also add a series of helper
functions that allow us to more easily deal with generics in pydantic.

* Require 3.11 for improved generic support